Rename app.py to sqlchain_app.py to distinguish it from future apps we will build.
